# frontend/mrsexam-ui/root/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "mrsexam-backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/FormBalbodhini": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "FormBalbodhini"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormBalbodhini"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormBalbodhini"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FormBalbodhini"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormBalbodhini/paged": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sessionNo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormBalbodhini/{id}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormBalbodhini"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormBalbodhini"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FormBalbodhini"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormBalbodhini/user/{userId}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormBalbodhini/centre/{centreCode}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "centreCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormBalbodhini/name/{name}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormBalbodhini/session/{sessionNo}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "sessionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/FormBalbodhini/region/{regionNo}": {
      "get": {
        "tags": [
          "FormBalbodhini"
        ],
        "parameters": [
          {
            "name": "regionNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FormBalbodhini": {
        "required": [
          "srNo"
        ],
        "type": "object",
        "properties": {
          "srNo": {
            "type": "integer",
            "format": "int32"
          },
          "rollNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "regionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 60,
            "type": "string",
            "nullable": true
          },
          "sessionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gender": {
            "maxLength": 1,
            "type": "string",
            "nullable": true
          },
          "formCentre": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "examCentre": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "marks1": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "totMarks": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "grade": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "undFlag": {
            "maxLength": 1,
            "type": "string",
            "nullable": true
          },
          "undClass": {
            "maxLength": 6,
            "type": "string",
            "nullable": true
          },
          "specialRank": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "result": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "dummyCentreCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "resgenRun": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "districtName": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "userId": {
            "maxLength": 15,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}